+++ tests/test_workflows.py
@@ -1,9 +1,11 @@
+from copy import deepcopy
+from random import shuffle
 from unittest import skip
 
-import pytest as pytest
-
 from pykechain.enums import StatusCategory, TransitionType, WorkflowCategory
-from pykechain.models.workflow import Workflow
+from pykechain.exceptions import NotFoundError
+from pykechain.models.input_checks import check_list_of_base
+from pykechain.models.workflow import Status, Workflow
 from tests.classes import TestBetamax
 
 
@@ -107,7 +109,20 @@ class TestWorkflowMethods(TestBetamax):
 
     def tearDown(self):
         super().tearDown()
-        self.workflow.delete()
+        try:
+            self.workflow.delete()
+        except NotFoundError:
+            # we pass this silently as workflow delete when the
+            # casettes are irrelevant and returns a 404 NotFoundError
+            pass
+
+    def test_edit_workflow_description(self):
+        new_description = "foo bar"
+        self.workflow.edit(
+            description=new_description
+        )
+        self.assertEqual(self.workflow.description, new_description)
+
 
     def test_workflow_in_and_activate(self):
         self.workflow.deactivate()
@@ -115,9 +130,13 @@ class TestWorkflowMethods(TestBetamax):
         self.workflow.activate()
         self.assertTrue(self.workflow.active)
 
-    # @skip("We are missing api endpoint to delete statusses directly as superuser.")
-    def test_create_workflow_transition(self):
-        """Test to create a transition on a workflow."""
+    def test_create_delete_status_and_create_delete_workflow_transition(self):
+        """Test to create a transition on a workflow.
+
+        - tests creation of status on workflow
+        - tests deletion of status
+        - tests deletion of transition on workflow
+        """
 
         count_transitions = len(self.workflow.transitions)
         test_status = self.workflow.create_status(
@@ -142,7 +161,31 @@ class TestWorkflowMethods(TestBetamax):
         test_status.delete()
         from_status.delete()
 
+    def test_workflow_status_order(self):
+        cloned_wf = self.workflow.clone(
+            name="___cloned_workflow_status_order", target_scope=self.workflow.scope_id
+        )
 
-class TestWorkflowTransitions(TestBetamax):
-    def test_create_workflow_transition(self):
-        ...
+        try:
+            status_ids = [s.id for s in deepcopy(cloned_wf.statuses)]
+            reversed_status_ids = list(reversed(status_ids))
+
+            cloned_wf.status_order = reversed_status_ids
+
+            self.assertListEqual(
+                reversed_status_ids, check_list_of_base(cloned_wf.status_order, Status)
+            )
+        finally:
+            cloned_wf.delete()
+
+    def test_workflow_unlink_and_link_transitions(self):
+        transitions = deepcopy(self.workflow.transitions)
+        the_transition_to_unlink = transitions[-1]
+
+        # unlink
+        self.workflow.unlink_transitions([the_transition_to_unlink])
+        self.assertListEqual(self.workflow.transitions, transitions[:-1])
+
+        # relink
+        self.workflow.link_transitions([the_transition_to_unlink])
+        self.assertListEqual(self.workflow.transitions, transitions)