+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ FORMS Feature
 * :+1: Added the `FormReferenceProperties` and tests (#1098)
 * :+1: Added the `ContextReferenceProperties` and tests (#1099)
 * :+1: Added 'Forms', 'Form models', 'Contexts' and 'Workflows' sidebar buttons in the `SideBarManager' and added ability to specify 'minimum_access_level' and 'alignment' per 'SideBarButton' (#1103)
+* :+1: Added the `Forms` class and tests (#1100)
 
 UNRELEASED
 ----------